refactor(iroh): restructure cli modules to match command structure
