+++ iroh/src/commands/author.rs
@@ -1,0 +1,55 @@
+use anyhow::{bail, Result};
+use clap::Parser;
+use futures::TryStreamExt;
+use iroh_base::base32::fmt_short;
+
+use iroh::client::quic::Iroh;
+use iroh_sync::AuthorId;
+
+use crate::config::ConsoleEnv;
+
+#[derive(Debug, Clone, Parser)]
+pub enum AuthorCommands {
+    /// Set the active author (only works within the Iroh console).
+    Switch { author: AuthorId },
+    /// Create a new author.
+    New {
+        /// Switch to the created author (only in the Iroh console).
+        #[clap(long)]
+        switch: bool,
+    },
+    /// List authors.
+    #[clap(alias = "ls")]
+    List,
+}
+
+impl AuthorCommands {
+    pub async fn run(self, iroh: &Iroh, env: &ConsoleEnv) -> Result<()> {
+        match self {
+            Self::Switch { author } => {
+                env.set_author(author)?;
+                println!("Active author is now {}", fmt_short(author.as_bytes()));
+            }
+            Self::List => {
+                let mut stream = iroh.authors.list().await?;
+                while let Some(author_id) = stream.try_next().await? {
+                    println!("{}", author_id);
+                }
+            }
+            Self::New { switch } => {
+                if switch && !env.is_console() {
+                    bail!("The --switch flag is only supported within the Iroh console.");
+                }
+
+                let author_id = iroh.authors.create().await?;
+                println!("{}", author_id);
+
+                if switch {
+                    env.set_author(author_id)?;
+                    println!("Active author is now {}", fmt_short(author_id.as_bytes()));
+                }
+            }
+        }
+        Ok(())
+    }
+}
